Reorganized contests components, fixed bug in invited contests stored procedure
