+++ client/src/Main.js
@@ -16,7 +16,7 @@ import SignUp from './components/SignUp/SignUp';
 import ContestDetails from './components/Contests/ContestDetails';
 import Profile from './components/Profile/Profile';
 import Users from './components/Users/Users';
-import ContestList from './components/ViewContestList/ContestList';
+import ContestList from './components/Contests/ContestList';
 import CreateContest from './components/CreateContest/CreateContest';
 import Dummy from './components/Dummy/FormDialog'
