--- client/src/components/ViewContestList/index.js
@@ -1,1 +1,0 @@
-export { default } from "./ContestList";
