+++ flask-api/data/Procedures.sql
@@ -134,7 +134,7 @@ Drop Procedure If Exists spGetInvitedContests;
 
 CREATE Procedure spGetInvitedContests (IN p_username varchar(64))
 BEGIN
-    SELECT contestID, contestName, description, startDate, endDate, status from Contest where contestID = (SELECT contestID from Contestuser where userID = (SELECT userID FROM Users WHERE username = p_username));
+    SELECT contestID, contestName, description, startDate, endDate, status from Contest where contestID in (SELECT contestID from Contestuser where userID = (SELECT userID FROM Users WHERE username = p_username));
 END //
 
 -- Get Contest Problems information
